Changed internal format for lincombs (for Finn): use now ONE BINARY FILE. OOPS: This change is not backward compatible with previous definitions. I also make a change that ensured that GRIDSUM is used to compute marginals for theta when int.strategy=grid and nhyper=1 (regardless of force_hessian_diagonal).

+++ gmrflib/io.h
@@ -97,8 +97,8 @@ int GMRFLib_io_seek(GMRFLib_io_tp * io,  size_t offset, int whence);
 int GMRFLib_io_read_next(GMRFLib_io_tp * io, void *ptr, const char *fmt);
 int GMRFLib_io_strip_blanks(char *line);
 int GMRFLib_sprintf(char **ptr, const char *fmt, ...);
-int GMRFLib_io_read(GMRFLib_io_tp * io, void *buf, unsigned int len);
-int GMRFLib_io_write(GMRFLib_io_tp * io, const void *buf, unsigned int len);
+int GMRFLib_io_read(GMRFLib_io_tp * io, void *buf, size_t len);
+int GMRFLib_io_write(GMRFLib_io_tp * io, const void *buf, size_t len);
 
 
 __END_DECLS
